+++ examples/blog/routes/posts/[uuid].tsx
@@ -1,6 +1,6 @@
 import { Handler, PageProps } from "$fresh/server.ts";
 import { Head } from "$fresh/runtime.ts";
-import { accepts } from "$std/http/mod.ts";
+import { respondWithObjectIfAcceptable } from "fedify/federation";
 import Comment from "../../components/Comment.tsx";
 import Post from "../../components/Post.tsx";
 import { federation } from "../../federation/mod.ts";
@@ -30,30 +30,11 @@ export const handler: Handler<PostPageData> = async (req, ctx) => {
   const post = await getPost(ctx.params.uuid);
   if (post == null) return await ctx.renderNotFound();
   const comments = await getComments(post.uuid);
-  const accept = accepts(
-    req,
-    "application/activity+json",
-    "application/ld+json",
-    "application/json",
-    "text/html",
-    "application/xhtml+xml",
-  );
-  if (
-    accept === "application/activity+json" ||
-    accept === "application/ld+json" || accept === "application/json"
-  ) {
-    const fedCtx = federation.createContext(req);
-    const article = toArticle(fedCtx, blog, post, comments);
-    const jsonLd = await article.toJsonLd(fedCtx);
-    return new Response(JSON.stringify(jsonLd), {
-      headers: {
-        "Content-Type": "application/activity+json",
-        Link:
-          `<${article.id}>; rel="alternate"; type="application/activity+json"`,
-        Vary: "Accept",
-      },
-    });
-  }
+  const fedCtx = federation.createContext(req);
+  const article = toArticle(fedCtx, blog, post, comments);
+  const response = await respondWithObjectIfAcceptable(article, req, fedCtx);
+  if (response != null) return response;
+
   const followers = await countFollowers();
   const data: PostPageData = {
     blog,