+++ federation/mod.ts
@@ -6,5 +6,10 @@
 export * from "./callback.ts";
 export * from "./collection.ts";
 export * from "./context.ts";
+export {
+  respondWithObject,
+  respondWithObjectIfAcceptable,
+  type RespondWithObjectOptions,
+} from "./handler.ts";
 export * from "./middleware.ts";
 export * from "./router.ts";